requirements.txt
asgiref==3.6.0
django-phonenumber-field==7.0.2
Django==4.2
pip==21.3.1
setuptools==60.2.0
sqlparse==0.4.3
tzdata==2023.3
wheel==0.37.1
phonenumbers==8.13.8
psycopg2-binary==2.9.6
